+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/MavenBuild.java
@@ -4,6 +4,18 @@ import java.io.IOException;
 
 public class MavenBuild extends Build {
 
+    /**
+     * This will be returned by the API in cases where no build has ever 
+     * been executed like {@link JobWithDetails#getLastBuild()} etc.
+     */
+    public static final MavenBuild BUILD_HAS_NEVER_RUN = new MavenBuild(-1, -1, "UNKNOWN");
+
+    private MavenBuild(int number, int queueId, String url) {
+        setNumber(number);
+        setQueueId(queueId);
+        setUrl(url);
+    }
+
     public MavenBuild() {
     }